Service settings and add routerlink
