+++ src/app/pages/account-settings/account-settings.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { SettingsService } from '../../services/settings.service';
 
 @Component({
   selector: 'app-account-settings',
@@ -6,38 +7,16 @@ import { Component, OnInit } from '@angular/core';
 })
 export class AccountSettingsComponent implements OnInit {
 
-   public linkTheme = document.querySelector('#theme');
-   public links!: NodeListOf<Element>;
-   constructor(){}
+   constructor(private settingsService: SettingsService){}
 
   ngOnInit(): void {
-    this.links = document.querySelectorAll('.selector');
-    this.checkCurrentTheme();
+    this.settingsService.checkCurrentTheme();
     
   }
   changeTheme(theme: string){
 
-    const url = `./assets/css/colors/${ theme }.css`;
-
-    this.linkTheme?.setAttribute('href', url);
-    localStorage.setItem('theme', url);
-
-    this.checkCurrentTheme();
+    this.settingsService.changeTheme(theme);
 
   }
 
-  checkCurrentTheme() {
-    
-    this.links.forEach( elem => {
-      elem.classList.remove('working');
-      const btnTheme = elem.getAttribute('data-theme');
-      const btnThemeUrl = `./assets/css/colors/${ btnTheme }.css`;
-      const currentTheme = this.linkTheme?.getAttribute('href');
-
-      if (btnThemeUrl === currentTheme ){
-        elem.classList.add('working');
-      }
-
-    });
-  }
 }
