+++ src/app/shared/sidebar/sidebar.component.html
@@ -13,7 +13,7 @@
                         <li><a href="javascript:void()">My Profile </a></li>
                         <li><a href="javascript:void()">My Balance</a></li>
                         <li><a href="javascript:void()">Inbox</a></li>
-                        <li><a href="javascript:void()">Account Setting</a></li>
+                        <li><a routerLink="account-settings">Account Setting</a></li>
                         <li><a href="javascript:void()">Logout</a></li>
                     </ul>
                 </li>
